test: share dispatch config harness
